+++ src/Generics/SOP/Classes.hs
@@ -40,6 +40,7 @@ module Generics.SOP.Classes
   , fn_3
   , fn_4
   , apFn_2
+  , Same
   , Prod
   , HAp(..)
     -- ** Derived functions
@@ -68,6 +69,12 @@ module Generics.SOP.Classes
     -- * Applying all injections
   , UnProd
   , HApInjs(..)
+    -- * Expanding sums to products
+  , HExpand(..)
+    -- * Transformation of index lists and coercions
+  , HTrans(..)
+  , hfromI
+  , htoI
   ) where
 
 #if !(MIN_VERSION_base(4,8,0))
@@ -149,6 +156,9 @@ fn   f = Fn $ \x -> f x
 fn_2 f = Fn $ \x -> Fn $ \x' -> f x x'
 fn_3 f = Fn $ \x -> Fn $ \x' -> Fn $ \x'' -> f x x' x''
 fn_4 f = Fn $ \x -> Fn $ \x' -> Fn $ \x'' -> Fn $ \x''' -> f x x' x'' x'''
+
+-- | Maps a structure to the same structure.
+type family Same (h :: (k1 -> *) -> (l1 -> *)) :: (k2 -> *) -> (l2 -> *)
 
 -- | Maps a structure containing sums to the corresponding
 -- product structure.
@@ -468,3 +478,110 @@ class (UnProd (Prod h) ~ h) => HApInjs (h :: (k -> *) -> (l -> *)) where
   -- @since 0.2.4.0
   --
   hapInjs :: (SListIN h xs) => Prod h f xs -> [h f xs]
+
+-- * Expanding sums to products
+
+-- | A class for expanding sum structures into corresponding product
+-- structures, filling in the slots not targeted by the sum with
+-- default values.
+--
+-- @since 0.2.5.0
+--
+class HExpand (h :: (k -> *) -> (l -> *)) where
+
+  -- | Expand a given sum structure into a corresponding product
+  -- structure by placing the value contained in the sum into the
+  -- corresponding position in the product, and using the given
+  -- default value for all other positions.
+  --
+  -- /Instances:/
+  --
+  -- @
+  -- 'hexpand', 'Generics.SOP.NS.expand_NS'  :: 'Generics.SOP.Sing.SListI' xs  => (forall x . f x) -> 'Generics.SOP.NS.NS'  f xs  -> 'Generics.SOP.NS.NP'  f xs
+  -- 'hexpand', 'Generics.SOP.NS.expand_SOP' :: 'SListI2' xss => (forall x . f x) -> 'Generics.SOP.NS.SOP' f xss -> 'Generics.SOP.NP.POP' f xss
+  -- @
+  --
+  -- /Examples:/
+  --
+  -- >>> hexpand Nothing (S (Z (Just 3))) :: NP Maybe '[Char, Int, Bool]
+  -- Nothing :* Just 3 :* Nothing :* Nil
+  -- >>> hexpand [] (SOP (S (Z ([1,2] :* "xyz" :* Nil)))) :: POP [] '[ '[Bool], '[Int, Char] ]
+  -- POP (([] :* Nil) :* ([1,2] :* "xyz" :* Nil) :* Nil)
+  --
+  -- @since 0.2.5.0
+  --
+  hexpand :: (SListIN (Prod h) xs) => (forall x . f x) -> h f xs -> Prod h f xs
+
+  -- | Variant of 'hexpand' that allows passing a constrained default.
+  --
+  -- /Instances:/
+  --
+  -- @
+  -- 'hcexpand', 'Generics.SOP.NS.cexpand_NS'  :: 'All'  c xs  => proxy c -> (forall x . c x => f x) -> 'Generics.SOP.NS.NS'  f xs  -> 'Generics.SOP.NP.NP'  f xs
+  -- 'hcexpand', 'Generics.SOP.NS.cexpand_SOP' :: 'All2' c xss => proxy c -> (forall x . c x => f x) -> 'Generics.SOP.NS.SOP' f xss -> 'Generics.SOP.NP.POP' f xss
+  -- @
+  --
+  -- /Examples:/
+  --
+  -- >>> hcexpand (Proxy :: Proxy Bounded) (I minBound) (S (Z (I 20))) :: NP I '[Bool, Int, Ordering]
+  -- I False :* I 20 :* I LT :* Nil
+  -- >>> hcexpand (Proxy :: Proxy Num) (I 0) (SOP (S (Z (I 1 :* I 2 :* Nil)))) :: POP I '[ '[Double], '[Int, Int] ]
+  -- POP ((I 0.0 :* Nil) :* (I 1 :* I 2 :* Nil) :* Nil)
+  --
+  -- @since 0.2.5.0
+  --
+  hcexpand :: (AllN (Prod h) c xs) => proxy c -> (forall x . c x => f x) -> h f xs -> Prod h f xs
+
+-- | A class for transforming structures into related structures with
+-- a different index list, as long as the index lists have the same shape
+-- and the elements and interpretation functions are suitably related.
+--
+-- @since 0.3.1.0
+--
+class (Same h1 ~ h2, Same h2 ~ h1) => HTrans (h1 :: (k1 -> *) -> (l1 -> *)) (h2 :: (k2 -> *) -> (l2 -> *)) where
+
+  -- | Transform a structure into a related structure given a conversion
+  -- function for the elements.
+  --
+  -- @since 0.3.1.0
+  --
+  htrans ::
+       AllZipN (Prod h1) c xs ys
+    => proxy c
+    -> (forall x y . c x y => f x -> g y)
+    -> h1 f xs -> h2 g ys
+
+  -- | Coerce a structure into a representationally equal structure.
+  --
+  -- /Examples:/
+  --
+  -- >>> hcoerce (I (Just LT) :* I (Just 'x') :* I (Just True) :* Nil) :: NP Maybe '[Ordering, Char, Bool]
+  -- Just LT :* (Just 'x' :* (Just True :* Nil))
+  -- >>> hcoerce (SOP (Z (K True :* K False :* Nil))) :: SOP I '[ '[Bool, Bool], '[Bool] ]
+  -- SOP (Z (I True :* (I False :* Nil)))
+  --
+  -- @since 0.3.1.0
+  hcoerce ::
+       (AllZipN (Prod h1) (LiftedCoercible f g) xs ys, HTrans h1 h2)
+    => h1 f xs -> h2 g ys
+
+-- | Specialization of 'hcoerce'.
+--
+-- @since 0.3.1.0
+--
+hfromI ::
+       (AllZipN (Prod h1) (LiftedCoercible I f) xs ys, HTrans h1 h2)
+    => h1 I xs -> h2 f ys
+hfromI = hcoerce
+
+-- | Specialization of 'hcoerce'.
+--
+-- @since 0.3.1.0
+--
+htoI ::
+       (AllZipN (Prod h1) (LiftedCoercible f I) xs ys, HTrans h1 h2)
+    => h1 f xs -> h2 I ys
+htoI = hcoerce
+
+-- $setup
+-- >>> import Generics.SOP